requirements.txt
gradio==4.44.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
mcp[cli]==1.9.3
